requirements.txt
linearfold-unofficial
linearpartition-unofficial
vaxpress
biopython
pandas
numpy
scikit-learn
snakemake